pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "domain-sentinel"
version = "0.1.0"
description = "Multi-site SSL, DNS, redirect, and endpoint health manager."
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
yaml = ["PyYAML>=6.0"]

[project.scripts]
domain-sentinel = "domain_sentinel.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
